requirements.txt
customtkinter==5.2.2
windnd>=1.0.7; sys_platform == "win32"
darkdetect==0.8.0
Deprecated==1.3.1
lxml==6.0.2
packaging==26.0
pikepdf==10.5.1
pillow==12.1.1
pypdf==6.9.1
wrapt==2.1.2
pytesseract
pdf2image
python-docx
PyPDF2
docx2pdf
pywin32>=306; sys_platform == 'win32'
openpyxl
reportlab
pdfplumber
pycryptodome
pdf2docx
